Add simple support for using Base classes
…  for application specific common use cases.

This removes the `Teckel::Config::default_constructor` and `Teckel::Config.results!` global settings in favour of parent classes.

Currently, Chains and Operations will inherit all the configurations of the parent class. That includes steps for Chains, which might or might not be useful. Since configurations can only be set once, a child class *cannot* overwrite a parents config (eg: `input` for Operations or `around` for Chains)