Move Operation success! and fail! logic into runner (#21)
* Move Operation success! and fail! logic into runner

The default throw-catch handling is now contained in the Runner class. This allows users to replace the entire runner logic.
The output and error values are now generated within the `success!` and `fail!` calls, before they are thrown. This allows for better stack traces, which now include the exact line in an Operation where an error or output contract was called with incorrect data.
Simple return values, that where considered success values, are now deprecated. This allows us to enforce users to be explicit about the return value. Also, providing a meaningful trace for an incorrectly called success contract is hard with plain return values.

* Remove support for simple return values

Any return value will just get discarded by the default runner.

* Update rubocop
* Fix dependencies for running example doc specs
* doc and bin helper for creating pages
* Spec pages doc examples on push